refactor: 상수 네이밍 변경 (magicNumber -> basket)
